Use forked raftbadger for log/stable stores
